Several changes to how binning works in interval_index_file.
First, the maximum has been increased (to 4096*1024*1024, bigger than the 32bit
integer we use to actaully store the positions can hold, thus MAX is actually 2**31).

Second, the number of levels of bins to use is determined from the max size passed
to create the index. Thus, where small intervals (e.g. contigs) previously the smallest
number of bins one could use had to include space for all the high level bins (
512 + 64 + 8 + 1 bins!). Now a small contig (under 128KB) has exactly one bin. Not
only does this save a ton of space, but it makes finding the bin for an interval much
faster.

The version number has been incremented, files with version < 2 will always use the
old binning scheme.

Impact: it is now tractable to index species such as

 possum: extremely large chromosomes were not supported before, now an
         additional bin will be added when creating indexes on large regions

 platypus: highly fragmented assembly with many small contigs, was taking a
           prohibitively long time to index, and wasting a ton of space, now
           can be indexed rapidly and compactly, e.g.:

		 2906252445 2007-07-11 11:02 ornAna1.maf
		   36021451 2007-07-11 13:22 ornAna1.maf.index